+++ openHAB/OpenHABSelectionTableViewController.swift
@@ -43,13 +43,14 @@ class OpenHABSelectionTableViewController: UITableViewController {
 
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell: UITableViewCell = tableView.dequeueReusableCell(withIdentifier: OpenHABSelectionTableViewController.tableViewCellIdentifier, for: indexPath)
-        let mapping = mappings[indexPath.row] as? OpenHABWidgetMapping
-        cell.textLabel?.text = mapping?.label
-        if selectionItem?.state == mapping?.command {
-            os_log("This item is selected", log: .viewCycle, type: .info)
-            cell.accessoryType = .checkmark
-        } else {
-            cell.accessoryType = .none
+        if let mapping = mappings[indexPath.row] as? OpenHABWidgetMapping {
+            cell.textLabel?.text = mapping.label
+            if selectionItem?.state == mapping.command {
+                os_log("This item is selected", log: .viewCycle, type: .info)
+                cell.accessoryType = .checkmark
+            } else {
+                cell.accessoryType = .none
+            }
         }
         return cell
     }